# internal/dto/staff.go
package dto

type Staff struct {
	ID           string 
	Name         string 
	Age          int    
	Gender       string 
	Email        string 
	Phone        string 
	Position     string 
	RestaurantID string 
}